+++ ResourceApp/ResourceApp/AppDelegate.swift
@@ -15,6 +15,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
   func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey : Any]?) -> Bool {
 
+    
 //    let defaults = UserDefaults.standard
 //    defaults.setValue(["nl"], forKey: "AppleLanguages")