+++ Sources/RswiftCore/Generators/ReuseIdentifierGenerator.swift
@@ -24,7 +24,7 @@ struct ReuseIdentifierStructGenerator: ExternalOnlyStructGenerator {
       .values
       .flatMap { $0.first }
 
-    let groupedReusables = deduplicatedReusables.groupedBySwiftIdentifier { $0.identifier }
+    let groupedReusables = deduplicatedReusables.grouped(bySwiftIdentifier: { $0.identifier })
     groupedReusables.printWarningsForDuplicatesAndEmpties(source: "reuseIdentifier", result: "reuseIdentifier")
 
     let reuseIdentifierProperties = groupedReusables
@@ -32,6 +32,7 @@ struct ReuseIdentifierStructGenerator: ExternalOnlyStructGenerator {
       .map { letFromReusable($0, at: externalAccessLevel) }
 
     return Struct(
+      availables: [],
       comments: ["This `\(qualifiedName)` struct is generated, and contains static references to \(reuseIdentifierProperties.count) reuse identifiers."],
       accessModifier: externalAccessLevel,
       type: Type(module: .host, name: structName),