+++ Sources/RswiftCore/Generators/SegueGenerator.swift
@@ -57,7 +57,7 @@ struct SegueStructGenerator: ExternalOnlyStructGenerator {
     var structs: [Struct] = []
 
     for (sourceType, seguesBySourceType) in deduplicatedSeguesWithInfo.grouped(by: { $0.sourceType }) {
-      let groupedSeguesWithInfo = seguesBySourceType.groupedBySwiftIdentifier { $0.segue.identifier }
+      let groupedSeguesWithInfo = seguesBySourceType.grouped(bySwiftIdentifier: { $0.segue.identifier })
 
       groupedSeguesWithInfo.printWarningsForDuplicatesAndEmpties(source: "segue", container: "for '\(sourceType)'", result: "segue")
 
@@ -71,6 +71,7 @@ struct SegueStructGenerator: ExternalOnlyStructGenerator {
     }
 
     return Struct(
+      availables: [],
       comments: ["This `\(qualifiedName)` struct is generated, and contains static references to \(structs.count) view controllers."],
       accessModifier: externalAccessLevel,
       type: Type(module: .host, name: structName),
@@ -130,6 +131,7 @@ struct SegueStructGenerator: ExternalOnlyStructGenerator {
 
     let functions = seguesWithInfoForSourceType.map { segueWithInfo -> Function in
       Function(
+        availables: [],
         comments: [
           "Optionally returns a typed version of segue `\(segueWithInfo.segue.identifier)`.",
           "Returns nil if either the segue identifier, the source, destination, or segue types don't match.",
@@ -153,6 +155,7 @@ struct SegueStructGenerator: ExternalOnlyStructGenerator {
     let typeName = SwiftIdentifier(name: sourceType.description)
 
     return Struct(
+      availables: [],
       comments: ["This struct is generated for `\(sourceType.name)`, and contains static references to \(properties.count) segues."],
       accessModifier: externalAccessLevel,
       type: Type(module: .host, name: typeName),