+++ Sources/RswiftCore/Generators/ValidatedStructGenerator.swift
@@ -29,6 +29,7 @@ class ValidatedStructGenerator: StructGenerator {
     }
 
     let validationStruct = Struct(
+      availables: [],
       comments: [],
       accessModifier: .filePrivate,
       type: Type(module: .host, name: "intern"),
@@ -37,6 +38,7 @@ class ValidatedStructGenerator: StructGenerator {
       properties: [],
       functions: [
         Function(
+          availables: [],
           comments: [],
           accessModifier: .filePrivate,
           isStatic: true,
@@ -85,6 +87,7 @@ private extension Struct {
     outputStruct.implements.append(TypePrinter(type: Type.Validatable))
     outputStruct.functions.append(
       Function(
+        availables: [],
         comments: [],
         accessModifier: externalAccessLevel,
         isStatic: true,