+++ Sources/RswiftCore/RswiftCore.swift
@@ -11,13 +11,9 @@ import Foundation
 import XcodeEdit
 
 public struct RswiftCore {
-  static var isEdgeEnabled = false
 
   static public func run(_ callInformation: CallInformation) throws {
-
     do {
-      RswiftCore.isEdgeEnabled = callInformation.edgeEnabled
-
       let xcodeproj = try Xcodeproj(url: callInformation.xcodeprojURL)
       let ignoreFile = (try? IgnoreFile(ignoreFileURL: callInformation.rswiftIgnoreURL)) ?? IgnoreFile()
 
@@ -28,9 +24,9 @@ public struct RswiftCore {
 
       let resources = Resources(resourceURLs: resourceURLs, fileManager: FileManager.default)
 
-      let generators: [StructGenerator] = [
+      var generators: [StructGenerator] = [
         ImageStructGenerator(assetFolders: resources.assetFolders, images: resources.images),
-        ColorStructGenerator(colorPalettes: resources.colors),
+        ColorStructGenerator(assetFolders: resources.assetFolders),
         FontStructGenerator(fonts: resources.fonts),
         SegueStructGenerator(storyboards: resources.storyboards),
         StoryboardStructGenerator(storyboards: resources.storyboards),
@@ -39,6 +35,14 @@ public struct RswiftCore {
         ResourceFileStructGenerator(resourceFiles: resources.resourceFiles),
         StringsStructGenerator(localizableStrings: resources.localizableStrings),
       ]
+
+      do {
+        let colorPaletteGenerator = ColorPaletteStructGenerator(palettes: resources.colors)
+        let colorPaletteGeneratorStruct = colorPaletteGenerator.generatedStructs(at: callInformation.accessLevel, prefix: "")
+        if !colorPaletteGeneratorStruct.externalStruct.isEmpty {
+          generators.append(colorPaletteGenerator)
+        }
+      }
 
       let aggregatedResult = AggregatedStructGenerator(subgenerators: generators)
         .generatedStructs(at: callInformation.accessLevel, prefix: "")