+++ Sources/RswiftCore/SwiftTypes/Struct.swift
@@ -10,6 +10,7 @@
 import Foundation
 
 struct Struct: UsedTypesProvider, SwiftCodeConverible {
+  let availables: [String]
   let comments: [String]
   let accessModifier: AccessLevel
   let type: Type
@@ -20,6 +21,13 @@ struct Struct: UsedTypesProvider, SwiftCodeConverible {
   var structs: [Struct]
   var classes: [Class]
 
+  var isEmpty: Bool {
+    return properties.isEmpty
+      && functions.isEmpty
+      && (structs.isEmpty || structs.all { $0.isEmpty })
+      && classes.isEmpty
+  }
+
   var usedTypes: [UsedType] {
     return [
         type.usedTypes,
@@ -28,11 +36,14 @@ struct Struct: UsedTypesProvider, SwiftCodeConverible {
         properties.flatMap(getUsedTypes),
         functions.flatMap(getUsedTypes),
         structs.flatMap(getUsedTypes),
-      ].flatten()
+      ]
+      .joined()
+      .array()
   }
 
   var swiftCode: String {
     let commentsString = comments.map { "/// \($0)\n" }.joined(separator: "")
+    let availablesString = availables.map { "@available(\($0))\n" }.joined(separator: "")
     let accessModifierString = accessModifier.swiftCode
     let implementsString = implements.count > 0 ? ": " + implements.map { $0.swiftCode }.joined(separator: ", ") : ""
 
@@ -71,6 +82,6 @@ struct Struct: UsedTypesProvider, SwiftCodeConverible {
     let bodyComponents = [typealiasString, varsString, functionsString, structsString, classesString, fileprivateInit].filter { $0 != "" }
     let bodyString = bodyComponents.joined(separator: "\n\n").indent(with: "  ")
 
-    return "\(commentsString)\(accessModifierString)struct \(type)\(implementsString) {\n\(bodyString)\n}"
+    return "\(commentsString)\(availablesString)\(accessModifierString)struct \(type)\(implementsString) {\n\(bodyString)\n}"
   }
 }