+++ Sources/RswiftCore/Util/UtilExtensions.swift
@@ -20,23 +20,16 @@ extension Array {
 // MARK: Sequence operations
 
 extension Sequence {
-  func grouped<Key>(by keySelector: (Iterator.Element) -> Key) -> [Key : [Iterator.Element]] {
-    var groupedBy = Dictionary<Key, [Iterator.Element]>()
-
-    for element in self {
-      let key = keySelector(element)
-      var array = groupedBy.removeValue(forKey: key) ?? []
-      array.append(element)
-      groupedBy[key] = array
-    }
-
-    return groupedBy
+  func grouped<Key>(by keyForValue: (Element) -> Key) -> [Key: [Element]] {
+    return Dictionary(grouping: self, by: keyForValue)
   }
-}
 
-extension Sequence where Iterator.Element : Sequence {
-  func flatten() -> [Iterator.Element.Iterator.Element] {
-    return flatMap { $0 }
+  func all(where predicate: (Element) throws -> Bool) rethrows -> Bool {
+    return !(try contains(where: { !(try predicate($0)) }))
+  }
+
+  func array() -> [Element] {
+    return Array(self)
   }
 }
 
@@ -46,13 +39,13 @@ extension String {
   var lowercaseFirstCharacter: String {
     if self.characters.count <= 1 { return self.lowercased() }
     let index = characters.index(startIndex, offsetBy: 1)
-    return substring(to: index).lowercased() + substring(from: index)
+    return self[..<index].lowercased() + self[index...]
   }
 
   var uppercaseFirstCharacter: String {
     if self.characters.count <= 1 { return self.uppercased() }
     let index = characters.index(startIndex, offsetBy: 1)
-    return substring(to: index).uppercased() + substring(from: index)
+    return self[..<index].uppercased() + self[index...]
   }
 
   func indent(with indentation: String) -> String {